% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator-class.R
\name{batchToDevice}
\alias{batchToDevice}
\title{sends a batch of data to device}
\usage{
batchToDevice(batch, device)
}
\arguments{
\item{batch}{the batch to send, usually a list of torch tensors}

\item{device}{which device to send batch to}
}
\value{
the batch on the required device
}
\description{
sends a batch of data to device
assumes batch includes lists of tensors to arbitrary nested depths
}
