% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{checkFileExists}
\alias{checkFileExists}
\title{helper function to check if a file exists}
\usage{
checkFileExists(file)
}
\arguments{
\item{file}{the file to check}
}
\description{
helper function to check if a file exists
}
