% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{checkHigher}
\alias{checkHigher}
\title{helper function to check that input is higher than a certain value}
\usage{
checkHigher(parameter, value)
}
\arguments{
\item{parameter}{the input parameter to check, can be a vector}

\item{value}{which value it should be higher than}
}
\description{
helper function to check that input is higher than a certain value
}
