% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{checkHigherEqual}
\alias{checkHigherEqual}
\title{helper function to check that input is higher or equal than a certain value}
\usage{
checkHigherEqual(parameter, value)
}
\arguments{
\item{parameter}{the input parameter to check, can be a vector}

\item{value}{which value it should be higher or equal than}
}
\description{
helper function to check that input is higher or equal than a certain value
}
