% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{checkIsClass}
\alias{checkIsClass}
\title{helper function to check class of input}
\usage{
checkIsClass(parameter, classes)
}
\arguments{
\item{parameter}{the input parameter to check}

\item{classes}{which classes it should belong to (one or more)}
}
\description{
helper function to check class of input
}
