% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepPatientLevelPrediction.R
\docType{package}
\name{DeepPatientLevelPrediction}
\alias{DeepPatientLevelPrediction-package}
\alias{DeepPatientLevelPrediction}
\title{DeepPatientLevelPrediction}
\description{
A package containing deep learning extensions for developing
prediction models using data in the OMOP CDM
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/DeepPatientLevelPrediction/}
  \item \url{https://github.com/OHDSI/DeepPatientLevelPrediction}
  \item Report bugs at \url{https://github.com/OHDSI/DeepPatientLevelPrediction/issues}
}

}
\author{
\strong{Maintainer}: Egill Fridgeirsson \email{e.fridgeirsson@erasmusmc.nl}

Authors:
\itemize{
  \item Egill Fridgeirsson \email{e.fridgeirsson@erasmusmc.nl}
  \item Jenna Reps \email{jreps@its.jnj.com}
  \item Seng Chan You
  \item Chungsoo Kim
  \item Henrik John
}

Other contributors:
\itemize{
  \item Observational Health Data Sciences and Informatics [copyright holder]
}

}
