% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Topologies.R
\name{doubleLayerNN}
\alias{doubleLayerNN}
\title{Double layer neural network}
\usage{
doubleLayerNN(inputN, layer1, layer2, outputN, layer_dropout)
}
\arguments{
\item{inputN}{Input neurons}

\item{layer1}{Layer 1 neurons}

\item{layer2}{Layer 2 neurons}

\item{outputN}{output neurons}

\item{layer_dropout}{layer_dropout to use}
}
\description{
Double layer neural network
}
