% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator-class.R
\name{EarlyStopping}
\alias{EarlyStopping}
\title{Earlystopping class}
\description{
Stops training if a loss or metric has stopped improving
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EarlyStopping-new}{\code{EarlyStopping$new()}}
\item \href{#method-EarlyStopping-call}{\code{EarlyStopping$call()}}
\item \href{#method-EarlyStopping-clone}{\code{EarlyStopping$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EarlyStopping-new"></a>}}
\if{latex}{\out{\hypertarget{method-EarlyStopping-new}{}}}
\subsection{Method \code{new()}}{
Creates a new earlystopping object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EarlyStopping$new(patience = 3, delta = 0, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{patience}}{Stop after this number of epochs if loss doesn't improve}

\item{\code{delta}}{How much does the loss need to improve to count as improvement}

\item{\code{verbose}}{If information should be printed out}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new earlystopping object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EarlyStopping-call"></a>}}
\if{latex}{\out{\hypertarget{method-EarlyStopping-call}{}}}
\subsection{Method \code{call()}}{
call the earlystopping object and increment a counter if loss is not
improving
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EarlyStopping$call(metric)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metric}}{the current metric value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EarlyStopping-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EarlyStopping-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EarlyStopping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
