% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator.R
\name{fitDeepPlpClassifier}
\alias{fitDeepPlpClassifier}
\title{fitDeepPlpClassifier}
\usage{
fitDeepPlpClassifier(
  trainData,
  modelSettings,
  hyperparameterSettings = NULL,
  analysisId,
  analysisPath,
  ...
)
}
\arguments{
\item{trainData}{the data to use}

\item{modelSettings}{modelSettings object}

\item{hyperparameterSettings}{PLP hyperparameter settings}

\item{analysisId}{Id of the analysis}

\item{analysisPath}{Path of the analysis}

\item{...}{Extra inputs}
}
\description{
Fits a deep learning estimator using PLP hyperparameter settings and
DeepPLP's cache-aware training loop.
}
