% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResNet_plp5.R
\name{fitResNet_plp5}
\alias{fitResNet_plp5}
\title{fitResNet_plp5}
\usage{
fitResNet_plp5(trainData, param, search = "grid", analysisId, ...)
}
\arguments{
\item{param}{parameters to use for model}

\item{...}{}

\item{population}{the study population dataframe}

\item{plpData}{plp data object}

\item{outcomeId}{Id of the outcome}

\item{cohortId}{Id of the cohort}
}
\description{
fits a ResNet model to data
}
