% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R, R/helpers.R
\name{listCols}
\alias{listCols}
\title{converts a sparse Matrix into a list of its columns, 
subsequently vapply can be used to apply functions over the list}
\usage{
listCols(m)

listCols(m)
}
\description{
converts a sparse Matrix into a list of its columns, 
subsequently vapply can be used to apply functions over the list
}
