% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predict_deepEstimator}
\alias{predict_deepEstimator}
\title{predict_deepEstimator}
\usage{
predict_deepEstimator(plpModel, population, plpData)
}
\arguments{
\item{plpModel}{The plpModel object}

\item{population}{Population dataframe}

\item{plpData}{plpData object}
}
\description{
prediction function for models using estimator class
}
