% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepNNTorch.R
\name{predictDeepNN}
\alias{predictDeepNN}
\title{Create predictions for a deep neural network}
\usage{
predictDeepNN(plpModel, data, cohort)
}
\arguments{
\item{plpModel}{The plpModel to predict for}

\item{data}{The data to make predictions for}

\item{cohort}{The cohort to use}
}
\description{
Create predictions for a deep neural network
}
