% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{py_to_r.polars.dataframe.frame.DataFrame}
\alias{py_to_r.polars.dataframe.frame.DataFrame}
\title{Use polars instead of pandas for default conversion from python to R}
\usage{
\method{py_to_r}{polars.dataframe.frame.DataFrame}(x)
}
\arguments{
\item{x}{A polars DataFrame object}
}
\value{
The same data.frame in R
}
\description{
Use polars instead of pandas for default conversion from python to R
}
