% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{r_to_py.data.frame}
\alias{r_to_py.data.frame}
\title{Use polars instead of pandas for default conversion}
\usage{
\method{r_to_py}{data.frame}(x, convert = FALSE)
}
\description{
Use polars instead of pandas for default conversion
}
