% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformer.R
\name{setCustomEmbeddingTransformer}
\alias{setCustomEmbeddingTransformer}
\title{Create default settings for a non-temporal transformer}
\usage{
setCustomEmbeddingTransformer(
  embeddingFilePath,
  estimatorSettings = setEstimator(learningRate = "auto", weightDecay = 1e-04, batchSize
    = 256, epochs = 2, seed = NULL, device = "cpu")
)
}
\arguments{
\item{embeddingFilePath}{path to the saved Poincare embedding}

\item{estimatorSettings}{created with `setEstimator`}
}
\description{
A transformer model with a Poincare embedding of diseases
}
\details{
Hierarchical embedding of disease concept in SNOMED medical terms
}
