% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepNNTorch.R
\name{setDeepNNTorch}
\alias{setDeepNNTorch}
\title{settings for a Deep neural network}
\usage{
setDeepNNTorch(
  units = list(c(128, 64), 128),
  layer_dropout = c(0.2),
  lr = c(1e-04),
  decay = c(1e-05),
  outcome_weight = c(1),
  batch_size = c(10000),
  epochs = c(100),
  device = "cpu",
  seed = NULL
)
}
\arguments{
\item{units}{A list of vectors for neurons per layer}

\item{layer_dropout}{Dropout to use per layer}

\item{lr}{Learning rate ot use}

\item{decay}{Weight decay to use}

\item{outcome_weight}{Weight for minority outcome in cost function}

\item{batch_size}{Batch size to use}

\item{epochs}{How many epochs to use}

\item{device}{Which device to use}

\item{seed}{A seed to make experiments more reproducible}
}
\description{
settings for a Deep neural network
}
