% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformer.R
\name{setDefaultTransformer}
\alias{setDefaultTransformer}
\title{Create default settings for a non-temporal transformer}
\usage{
setDefaultTransformer(
  estimatorSettings = setEstimator(learningRate = "auto", weightDecay = 1e-04, batchSize
    = 512, epochs = 10, seed = NULL, device = "cpu")
)
}
\arguments{
\item{estimatorSettings}{Estimator settings created by \code{\link[=setEstimator]{setEstimator()}}.}
}
\value{
A \code{modelSettings} object for use with \code{PatientLevelPrediction}.
}
\description{
Creates settings for the package's default non-temporal transformer model.
}
\details{
The architecture and default hyperparameters are based on
\href{https://arxiv.org/abs/2106.11959}{Gorishniy et al. (2021)}.
}
\examples{
transformerSettings <- setDefaultTransformer()
transformerSettings$param[[1]]$numBlocks

}
