% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Topologies.R
\name{singleLayerNN}
\alias{singleLayerNN}
\title{A single layer neural network}
\usage{
singleLayerNN(inputN, layer1, outputN = 2, layer_dropout)
}
\arguments{
\item{inputN}{Input neurons}

\item{layer1}{Layer 1 neurons}

\item{outputN}{Output neurons}

\item{layer_dropout}{Layer dropout to use}
}
\description{
A single layer neural network
}
