% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{snakeCaseToCamelCase}
\alias{snakeCaseToCamelCase}
\title{Convert a camel case string to snake case}
\usage{
snakeCaseToCamelCase(string)
}
\arguments{
\item{string}{The string to be converted}
}
\value{
A string
}
\description{
Convert a camel case string to snake case
}
