% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepPatientLevelPrediction.R
\name{torch}
\alias{torch}
\title{PyTorch Module}
\format{
An object of class \code{python.builtin.module}
}
\value{
The \code{torch} Python module.
}
\description{
The \code{torch} module object is the equivalent of
\code{reticulate::import("torch")} and is provided mainly as a convenience.
Accessing the module initializes Python and requires the Python dependencies
listed in \code{SystemRequirements} in the package \code{DESCRIPTION} file.
}
\examples{
\dontrun{
# Requires the Python dependencies described in vignette("Installing").
torch$randn(10L)
}
}
