% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Topologies.R
\name{tripleLayerNN}
\alias{tripleLayerNN}
\title{Triple layer neural network}
\usage{
tripleLayerNN(inputN, layer1, layer2, layer3, outputN, layer_dropout)
}
\arguments{
\item{inputN}{Input neurons}

\item{layer1}{amount of layer 1 neurons}

\item{layer2}{amount of layer 2 neurons}

\item{layer3}{amount of layer 3 neurons}

\item{outputN}{Number of output neurons}

\item{layer_dropout}{The dropout to use in layer}
}
\description{
Triple layer neural network
}
