% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{Dataset}
\alias{Dataset}
\title{A torch dataset}
\usage{
Dataset(data, labels = NULL, numericalIndex = NULL)
}
\arguments{
\item{data}{a dataframe like object with the covariates}

\item{labels}{a dataframe with the labels}

\item{numericalIndex}{in what column numeric data is in (if any)}
}
\description{
A torch dataset
}
