% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepPatientLevelPrediction.R
\docType{package}
\name{DeepPatientLevelPrediction}
\alias{DeepPatientLevelPrediction}
\title{DeepPatientLevelPrediction}
\description{
A package for running predictions using deep learning on
data in the OMOP CDM
}
