% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepPatientLevelPrediction.R
\docType{package}
\name{DeepPatientLevelPrediction}
\alias{DeepPatientLevelPrediction}
\title{DeepPatientLevelPrediction}
\description{
A package containing deep learning extensions for developing
prediction models using data in the OMOP CDM
}
