% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepNNTorch.R
\name{fitDeepNNTorch}
\alias{fitDeepNNTorch}
\title{Fits a deep neural network}
\usage{
fitDeepNNTorch(trainData, modelSettings, search = "grid", analysisId)
}
\arguments{
\item{trainData}{Training data object}

\item{modelSettings}{modelSettings object}

\item{search}{Which kind of search strategy to use}

\item{analysisId}{Analysis Id}
}
\description{
Fits a deep neural network
}
