% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator.R
\name{fitEstimator}
\alias{fitEstimator}
\title{fitEstimator}
\usage{
fitEstimator(trainData, param, analysisId, ...)
}
\arguments{
\item{trainData}{the data to use}

\item{param}{model parameters}

\item{analysisId}{Id of the analysis}

\item{...}{}
}
\description{
fits a deep learning estimator to data.
}
