% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator.R
\name{fitEstimator}
\alias{fitEstimator}
\title{fitEstimator}
\usage{
fitEstimator(trainData, modelSettings, analysisId, analysisPath, ...)
}
\arguments{
\item{trainData}{the data to use}

\item{modelSettings}{modelSettings object}

\item{analysisId}{Id of the analysis}

\item{analysisPath}{Path of the analysis}

\item{...}{Extra inputs}
}
\description{
fits a deep learning estimator to data.
}
