% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator.R
\name{gridCvDeep}
\alias{gridCvDeep}
\title{gridCvDeep}
\usage{
gridCvDeep(mappedData, labels, modelSettings, modelLocation)
}
\arguments{
\item{mappedData}{Mapped data with covariates}

\item{labels}{Dataframe with the outcomes}

\item{modelSettings}{Settings of the model}

\item{modelLocation}{Where to save the model}
}
\description{
Performs grid search for a deep learning estimator
}
