% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator.R
\name{gridCvDeep}
\alias{gridCvDeep}
\title{gr idCvDeep}
\usage{
gridCvDeep(matrixData, labels, settings, modelLocation, paramSearch)
}
\arguments{
\item{matrixData}{Data in sparse matrix format}

\item{labels}{Dataframe with the outcomes}

\item{settings}{Settings of the model}

\item{modelLocation}{Where to save the model}

\item{paramSearch}{model parameters to perform search over}
}
\description{
Performs grid search for a deep learning estimator
}
