% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResNet_plp5.R
\name{predictDeepEstimator}
\alias{predictDeepEstimator}
\title{predictDeepEstimator}
\usage{
predictDeepEstimator(plpModel, data, cohort)
}
\arguments{
\item{plpModel}{the plpModel}

\item{data}{plp data object or a torch dataset}

\item{cohort}{a data.frame with the rowIds of the people to predict risk for}
}
\description{
the prediction function for the binary classification deep learning models
}
