% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomEmbeddingModel.R
\name{setCustomEmbeddingModel}
\alias{setCustomEmbeddingModel}
\title{Create default settings a model using custom embeddings}
\usage{
setCustomEmbeddingModel(
  embeddingFilePath,
  estimatorSettings = setEstimator(learningRate = "auto", weightDecay = 1e-04, batchSize
    = 256, epochs = 2, seed = NULL, device = "cpu"),
  modelSettings = setTransformer(numBlocks = 3, dimToken = 16, dimOut = 1, numHeads = 4,
    attDropout = 0.2, ffnDropout = 0.1, resDropout = 0, dimHidden = 32, estimatorSettings
    = estimatorSettings, hyperParamSearch = "random", randomSample = 1)
)
}
\arguments{
\item{embeddingFilePath}{path to the saved Poincare embedding}

\item{estimatorSettings}{created with `setEstimator`}

\item{modelSettings}{for the model to use, needs to have an embedding layer}
}
\value{
settings for a model using custom embeddings:w
}
\description{
A model that uses custom embeddings such as Poincare embeddings
}
