% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResNet.R
\name{setDefaultResNet}
\alias{setDefaultResNet}
\title{setDefaultResNet}
\usage{
setDefaultResNet(device = "cpu", batchSize = 1024, epochs = 10, seed = NULL)
}
\arguments{
\item{device}{Which device to run analysis on, either 'cpu' or 'cuda', default: 'cpu'}

\item{batchSize}{Size of batch, default: 1024}

\item{epochs}{Number of epochs to run, default: 10}

\item{seed}{Random seed to use}
}
\description{
Creates settings for a default ResNet model
}
\details{
Model architecture from by https://arxiv.org/abs/2106.11959 . 
Hyperparameters chosen by a experience on a few prediction problems.
}
