% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResNet.R
\name{setDefaultResNet}
\alias{setDefaultResNet}
\title{setDefaultResNet}
\usage{
setDefaultResNet(
  estimatorSettings = setEstimator(learningRate = "auto", weightDecay = 1e-06, device =
    "cpu", batchSize = 1024L, epochs = 50L, seed = NULL)
)
}
\arguments{
\item{estimatorSettings}{created with ```setEstimator```}
}
\description{
Creates settings for a default ResNet model
}
\details{
Model architecture from by https://arxiv.org/abs/2106.11959 . 
Hyperparameters chosen by a experience on a few prediction problems.
}
