% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformer.R
\name{setDefaultTransformer}
\alias{setDefaultTransformer}
\title{Create default settings for a non-temporal transformer}
\usage{
setDefaultTransformer(
  estimatorSettings = setEstimator(learningRate = "auto", weightDecay = 1e-04, batchSize
    = 512, epochs = 10, seed = NULL, device = "cpu")
)
}
\arguments{
\item{estimatorSettings}{created with `setEstimator`}
}
\description{
A transformer model with default hyperparameters
}
\details{
from https://arxiv.org/abs/2106.11959
Default hyperparameters from paper
}
