% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformer.R
\name{setDefaultTransformer}
\alias{setDefaultTransformer}
\title{Create default settings for a non-temporal transformer}
\usage{
setDefaultTransformer(
  device = "cpu",
  batchSize = 512,
  epochs = 10,
  seed = NULL
)
}
\arguments{
\item{device}{Which device to run analysis on, either 'cpu' or 'cuda', default: 'cpu'}

\item{batchSize}{Size of batch, default: 512}

\item{epochs}{Number of epochs to run, default: 10}

\item{seed}{random seed to use}
}
\description{
A transformer model with default hyperparameters
}
\details{
from https://arxiv.org/abs/2106.11959
Default hyperparameters from paper
}
