% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator.R
\name{setEstimator}
\alias{setEstimator}
\title{setEstimator}
\arguments{
\item{learningRate}{what learning rate to use}

\item{weightDecay}{what weight_decay to use}

\item{optimizer}{which optimizer to use}

\item{scheduler}{which learning rate scheduler to use}

\item{criterion}{loss function to use}

\item{earlyStopping}{If earlyStopping should be used which stops the training of your metric is not improving}

\item{earlyStoppingMetric}{Which parameter to use for early stopping}

\item{patience}{patience for earlyStopper}

\item{hyperparameterMetric}{which metric to use for hyperparameter, loss, auc, auprc or a custom function}
}
\description{
creates settings for the Estimator, which takes a model and trains it
}
