% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransferLearning.R
\name{setFinetuner}
\alias{setFinetuner}
\title{setFinetuner}
\usage{
setFinetuner(
  modelPath,
  estimatorSettings = setEstimator(learningRate = learningRate, weightDecay =
    weightDecay, batchSize = batchSize, epochs = epochs, device = device, optimizer =
    optimizer, scheduler = scheduler, criterion = criterion, earlyStopping =
    earlyStopping, metric = metric, seed = seed)
)
}
\arguments{
\item{modelPath}{path to existing plpModel directory}

\item{estimatorSettings}{settings created with `setEstimator`}
}
\description{
creates settings for using transfer learning to finetune a model
}
