% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLayerPerceptron.R
\name{setMultiLayerPerceptron}
\alias{setMultiLayerPerceptron}
\title{setMultiLayerPerceptron}
\usage{
setMultiLayerPerceptron(
  numLayers = c(1:8),
  sizeHidden = c(2^(6:9)),
  dropout = c(seq(0, 0.3, 0.05)),
  sizeEmbedding = c(2^(6:9)),
  estimatorSettings = setEstimator(learningRate = "auto", weightDecay = c(1e-06, 0.001),
    batchSize = 1024, epochs = 30, device = "cpu"),
  hyperParamSearch = "random",
  randomSample = 100,
  randomSampleSeed = NULL
)
}
\arguments{
\item{numLayers}{Number of layers in network, default: 1:8}

\item{sizeHidden}{Amount of neurons in each default layer,
default: 2^(6:9) (64 to 512)}

\item{dropout}{How much dropout to apply after first linear,
default: seq(0, 0.3, 0.05)}

\item{sizeEmbedding}{Size of embedding default: 2^(6:9) (64 to 512)}

\item{estimatorSettings}{settings of Estimator created with `setEstimator`}

\item{hyperParamSearch}{Deprecated. Use PLP `hyperparameterSettings`
instead.}

\item{randomSample}{Deprecated. Use PLP `hyperparameterSettings` instead.}

\item{randomSampleSeed}{Deprecated. Use PLP `hyperparameterSettings`
instead.}
}
\description{
Creates settings for a Multilayer perceptron model
}
\details{
Model architecture
}
