% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResNet.R
\name{setResNet}
\alias{setResNet}
\title{setResNet}
\usage{
setResNet(
  numLayers = as.integer(1:8),
  sizeHidden = as.integer(2^(6:10)),
  hiddenFactor = as.integer(1:4),
  residualDropout = c(seq(0, 0.5, 0.05)),
  hiddenDropout = c(seq(0, 0.5, 0.05)),
  sizeEmbedding = as.integer(2^(6:9)),
  estimatorSettings = setEstimator(learningRate = "auto", weightDecay = c(1e-06, 0.001),
    device = "cpu", batchSize = 1024L, epochs = 30L, seed = NULL),
  hyperParamSearch = "random",
  randomSample = 100,
  randomSampleSeed = NULL
)
}
\arguments{
\item{numLayers}{Number of layers in network, default: 1:16}

\item{sizeHidden}{Amount of neurons in each default layer, default: 2^(6:10) (64 to 1024)}

\item{hiddenFactor}{How much to grow the amount of neurons in each ResLayer, default: 1:4}

\item{residualDropout}{How much dropout to apply after last linear layer in ResLayer, default: seq(0, 0.3, 0.05)}

\item{hiddenDropout}{How much dropout to apply after first linear layer in ResLayer, default: seq(0, 0.3, 0.05)}

\item{sizeEmbedding}{Size of embedding layer, default: 2^(6:9) (64 to 512)}

\item{estimatorSettings}{created with ```setEstimator```}

\item{hyperParamSearch}{Which kind of hyperparameter search to use random sampling or exhaustive grid search. default: 'random'}

\item{randomSample}{How many random samples from hyperparameter space to use}

\item{randomSampleSeed}{Random seed to sample hyperparameter combinations}
}
\description{
Creates settings for a ResNet model
}
\details{
Model architecture from by https://arxiv.org/abs/2106.11959
}
