% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformer.R
\name{setTransformer}
\alias{setTransformer}
\title{setTransformer}
\usage{
setTransformer(
  numBlocks = 3,
  dimToken = 96,
  dimOut = 1,
  numHeads = 8,
  attDropout = 0.25,
  ffnDropout = 0.25,
  resDropout = 0,
  dimHidden = 512,
  weightDecay = 1e-06,
  learningRate = 3e-04,
  batchSize = 1024,
  epochs = 10,
  device = "cpu",
  hyperParamSearch = "random",
  randomSamples = 100,
  seed = NULL
)
}
\arguments{
\item{numBlocks}{TODO: add desc}

\item{dimToken}{TODO: add desc}

\item{dimOut}{TODO: add desc}

\item{numHeads}{TODO: add desc}

\item{attDropout}{TODO: add desc}

\item{ffnDropout}{TODO: add desc}

\item{resDropout}{TODO: add desc}

\item{dimHidden}{TODO: add desc}

\item{weightDecay}{TODO: add desc}

\item{learningRate}{TODO: add desc}

\item{batchSize}{TODO: add desc}

\item{epochs}{TODO: add desc}

\item{device}{TODO: add desc}

\item{hyperParamSearch}{TODO: add desc}

\item{randomSamples}{TODO: add desc}

\item{seed}{TODO: add desc}
}
\description{
A transformer model
}
\details{
from https://arxiv.org/abs/2106.11959
}
