% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformer.R
\name{setTransformer}
\alias{setTransformer}
\title{create settings for training a transformer}
\usage{
setTransformer(
  numBlocks = 3,
  dimToken = 192,
  dimOut = 1,
  numHeads = 8,
  attDropout = 0.2,
  ffnDropout = 0.1,
  dimHidden = 256,
  dimHiddenRatio = NULL,
  temporal = FALSE,
  useRope = FALSE,
  maxSequenceLength = 256,
  truncation = "tail",
  estimatorSettings = setEstimator(weightDecay = 1e-06, batchSize = 1024, epochs = 10,
    seed = NULL),
  hyperParamSearch = "random",
  randomSample = 1,
  randomSampleSeed = NULL
)
}
\arguments{
\item{numBlocks}{number of transformer blocks}

\item{dimToken}{dimension of each token (embedding size)}

\item{dimOut}{dimension of output, usually 1 for binary
problems}

\item{numHeads}{number of attention heads}

\item{attDropout}{dropout to use on attentions}

\item{ffnDropout}{dropout to use in feedforward block}

\item{dimHidden}{dimension of the feedworward block}

\item{dimHiddenRatio}{dimension of the feedforward block as a ratio
of dimToken (embedding size)}

\item{temporal}{Whether to use a transformer with temporal data}

\item{useRope}{Whether to use ROPE (Relative Positional Encoding)}

\item{maxSequenceLength}{Maximum sequence length, sequences longer than
this will be truncated and/or padded to this length}

\item{truncation}{Truncation method, only 'tail' is supported}

\item{estimatorSettings}{created with `setEstimator`}

\item{hyperParamSearch}{what kind of hyperparameter search to do,
default 'random'}

\item{randomSample}{How many samples to use in hyperparameter
search if random}

\item{randomSampleSeed}{Random seed to sample hyperparameter
combinations}
}
\value{
list of settings for the transformer model
}
\description{
A transformer model
}
\details{
The non-temporal transformer is from https://arxiv.org/abs/2106.11959
}
