% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepPatientLevelPrediction.R
\docType{data}
\name{torch}
\alias{torch}
\title{Pytorch module}
\format{
An object of class `python.builtin.module`
}
\value{
the torch Python module
}
\description{
The `torch` module object is the equivalent of
`reticulate::import("torch")` and provided mainly as a convenience.
}
\keyword{datasets}
